refactor(labs): Moved space function into _labs/core
BREAKING CHANGE: Space function is now in _labs/core
