+++ modules/_labs/core/react/index.ts
@@ -1,5 +1,6 @@
 import type from './lib/type';
+import space from './lib/space';
 
 export default type;
-export {type};
+export {type, space};
 export * from './lib/type';
