+++ modules/_labs/core/react/package.json
@@ -38,6 +38,7 @@
   },
   "dependencies": {
     "@workday/canvas-colors-web": "^0.17.12",
-    "@workday/canvas-kit-react-core": "^3.0.0-alpha.7"
+    "@workday/canvas-kit-react-core": "^3.0.0-alpha.7",
+    "emotion": "^9.2.12"
   }
 }
