+++ modules/core/react/stories/stories.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import styled, {css} from 'react-emotion';
 import withReadme from 'storybook-readme/with-readme';
 
-import canvas, {H1, space, spacing} from '..';
+import canvas, {H1} from '..';
 import README from '../README.md';
 
 export const inverseStyle = {
@@ -69,19 +69,6 @@ export const type = (hierarchy: any) => (
 
 storiesOf('Core', module)
   .addDecorator(withReadme(README))
-  .add('Space', () => {
-    const Box = styled('div')(space);
-
-    return (
-      <div className="story">
-        <div>
-          <Box style={{border: '1px solid #eee'}} p={spacing.xl} pb={64} m={40} mx={10}>
-            Margin and Padding
-          </Box>
-        </div>
-      </div>
-    );
-  })
   .add('Depth', () => {
     const Card = styled('div')({
       width: 200,
